Output attributes in snake case.
Summary:
Users are always accustomed to attributes being in snake case,
not camel case. They're camel case in the query output at the
moment because that's what the marshaller uses. This converts
the output and the arguments to `--output-attributes` to be
snake case.
Note `attrfilter` already uses snake case.

Test Plan: CI