[halide] Add halide_library rule
Summary:
Halide (http://halide-lang.org) is a "programming language designed to make it
easier to write high-performance image processing code".

Halide pipelines are typically JIT-compiled, which requires no special support
from Buck, but which requires that the application link to the Halide library
(libHalide.so). For use cases where this is undesirable, Halide also supports
"static compilation", which requires building and running a "compiler" program
as part of the build, the output of which is the object file that must be
linked into the final program (https://github.com/halide/Halide/wiki/Static-vs.-JIT-compilation).
This diff adds support for "static" compilation of Halide pipelines to Buck.

This rule is intended be used like this:

  halide_library(
    # Your library name.
    name = 'brighter',

    # Your pipeline + compiler sources.
    srcs = ['halide/main.cpp'],

    # Any dependencies for your compiler. Note that targets that depend on
    # this rule WILL NOT include or link the output(s) of these targets.
    compiler_deps = [
      '//Vendor/Boost:boost-filesystem',
      '//Vendor/Boost:boost-program-options',
      '//VendorLib/Halide:halide'
    ],
  )

  cxx_library(
    name = 'CreativeLibraryCore',
    srcs = [ 'some_file.cpp' ],
    deps = [
      # ... other deps
      ':brighter',
    ],
    visibility = ['PUBLIC'],
  )

Test Plan:
- Built and ran an ios_binary target that depends on a halide_library target in the simulator. Checked that the Halide pipeline code was compiled and linked correctly.
- Ran ##docs/soyweb_local.sh##, browsed to localhost:9811 and checked that the ##halide_library## documentation rendered correctly.