[halide] Allow direct update of build script phases for project
Summary:
For halide support, I'm adding a PBXNativeTarget to the project that will have
only a "Run Script" phase (it won't have any sources or dependencies). I need
to be able to mutate the preBuildRunScriptPhases directly to do this. This
diff exposes a new interface from the NewNativeTargetProjectMutator to allow
callers to do that.

Test Plan: Used the new methods in the next diff in the stack.