Don't create directories based on unflavoured targets.
Summary:
Unflavoured build targets aren't guaranteed to be unique across
build rules so it's dangerous to create scratch and output
directories based on them. This does not change the directories
that are being created, but it does remove the methods to make
it harder to introduce new instances of this issue.

Test Plan: CI