Fix phantom root issue on Windows
Summary:
On Windows, it is possible that the system will have a drive for something that does not
exist such as a floppy disk or SD card.  The drive exists, but it does not contain
anything useful, so Buck should not consider it as a cacheable location.

Test Plan: I can finally run Buck and Buck's tests on my Surface Pro 3 again