[halide] Basic Xcode support for halide filters
Summary:
This diff adds support for building halide filters during project generation,
and linking the resulting generated object code. We support building and
running halide pipelines both in the simulator and on device. The selection
of the target platform is done based on the ##ARCHS## environment
variable set by Xcode.

Test Plan:
- Ran 'buck project' for a target that depends transitively on a halide
  filter; opened the project, built the app in Xcode and ran it in the
  simulator and on an iOS device.