[Test] Skip JNI test for MacOS.
Summary:
The assumption for the JNI folder at java home only works reliably for
linux systems. For example, for MacOS, some of the headers are under
`$JAVA_HOME/include/darwin/` instead of `$JAVA_HOME/include/linux`. This
change set change the test to only run for Linux to make the MacOS test
pass. We need to find a reliable way to find the headers on different OSes.

Test:
`ant clean lint && ant && buck test`