Relocate doc comment about input based rule keys and add cross references.
Summary:
I think it makes sense to describe what input based rule keys are in the
interface that a rule writer is going to use, rather than in the machinery that
implements it.

To that end I moved the documentation describing input based rule keys into the
tagging interface, and added cross references to both the interface and the
implementation.

Test Plan: visual scrutiny.