Allow NDK CPU ABIs to be disabled in buckconfig
Summary:
An ndk_library that depends on a cxx_library will end up depending on
all architectures of that cxx_library, even if we're building an app
that uses CPU filters.  That can be problematic for libraries that can't
be built for certain architectures.

This gives us the option to globally disable some NDK platforms so we
never try to build for one that we don't support.

Test Plan: One sec.