Allow to remove config values by passing empty value via --config
Summary:
We have a way to override the config value via `--config section.property=value`, but we don't have the way to
remove the value. These changes add ability to use `--config section.property=` to perform the removal of the config
value, as it would never set in the buck config file.

Test Plan: Unit Tests