Fix bug in build-with-buck with library targets
Summary:
Running `buck project //src:foo --build-with-buck` creates a project that cannot be built in Xcode if //src:foo is a library target. Previously we used one shell script that 1) built the project with Buck and 2) fixed UUIDs in the binary. Now we add the shell script to build in all cases, but only add the fix_uuid script if it is an apple_bundle.

Closes https://github.com/facebook/buck/pull/497

Test Plan:
Unit tests updated. `ant java-test -Dtest.class=ProjectGeneratorTest`. Also, build an `apple_bundle`, `apple_binary` and `apple_library`, with `buck project <target> --build-with-buck --combined-project` and make sure all generated Xcode projects build correctly.