Treat stringsdict files as binary so that Xcode will convert to binary plist when copying them
Summary:
`.stringsdict` files are used by iOS for localization, and are just `.plist` files with a different extension.  Right now, we just copy them without conversion which contributes to unneeded app size.

Set the file type to `file.bplist` which will make Xcode do the conversion to binary.

We do this already for Buck builds (see `addResourceProcessingSteps()` in `AppleBundle`), this enables it for Xcode builds.

Test Plan: Test in Xcode with projects containing `.stringsdict`, build in Xcode, see in the logs that the `.stringsdict` files are converted to binary.