[halide] Fix missing header issue when building in Xcode
Summary:
Building in Xcode sometimes fails when the project includes a halide_library
target, because Xcode can attempt to build the dependent sources before running
the Halide "compiler"; this fails because the generated header for the library
is missing. To fix this, we run the "compiler" once at 'buck project' time for
each halide_library in the project, to create the generated header.

Test Plan:
- Did 'buck project' on a test application target which transitively depends on
  a halide_library. Opened the project in Xcode, built the app, and checked that
  the build succeeded, even if the halide target was built out-of-order.