[RFC] Refactor common rule key factory code into abstract class
Summary: The `DefaultRuleKeyBuilderFactory` contains the code which does the
reflective walk of the build rule and adds default properties to the
rule key (e.g. buck version).  Since we want this stuff for all our
rule keys, all the other rule key factories end up sub-classing this
to get this behavior.

However, it's useful for the other rule key factories to use their
own special sub-classes of `RuleKeyBuilder` (e.g. to capture all inputs
added to the rule key).  Since the `DefaultRuleKeyBuilderFactory`s
`newInstance` method returns a `RuleKeyBuilder` type, it makes it
difficult for the other factory methods to to call that and then use
the any added methods in its' `RuleKeyBuilder` sub-class.

This diff introduces a new `ReflectiveRuleKeyBuilderFactory` abstract
class and moves all the reflactive and default property code there
(in the `newInstance` method).  This tries to add value in two ways:
- The `ReflectiveRuleKeyBuilderFactory` uses generics so that the
  actual `RuleKeyBuilder` type is preserved.  This means the other
  factory types can introduce new `RuleKeyBuilder` sub-classes and
  still re-used the `newInstance` method from the this super-clase.
- The subsequent refactoring of the rule key factories makes them a
  bit simpler, replacing `Function` objects with overrides of the
  `newInstance` method which adds in needed additional factory-specific
  settings (this also fixes the existing code so that additional
  settings are added *after* the default ones).

Test Plan: CI

Reviewed By: Coneko

fb-gh-sync-id: 9c2fc25