Output last N lines of debug logs from test to console if test fails
Summary: Now that Robolectric and Apple tests can write their
debug logs to a file, this diff adds the ability to include
the logs in console output when a test fails.

I set N to 50 to start. We can make this configurable either via
`.buckconfig` or via the CLI.

Next step is to wire this up to the Java and Apple test rules.

Test Plan: Unit tests updated. `ant java-test -Dtest.class=TestResultFormatterTest`

Reviewed By: shs96c

fb-gh-sync-id: fff71ea