Always make the unflavoured target nodes addressable.
Summary: The target graph contains flavoured nodes, however when graph
enhancing because flavours are added and removed, it can be
very hard to track down the target node that originally created
a certain build rule. This adds the unflavoured target node to
the index. Note this is not guaranteed to be the equal to the
flavoured target node, as implicit deps inferring can change
the dependencies based on the flavours.

Test Plan: CI

Reviewed By: beefon

fb-gh-sync-id: 53cc2a9