Work around ar issue on Apple platforms where appending a .o file to an existing archive would overwrite any existing .o of the same name
Summary: When we split a long `ar` invocation into multiple commands via `CommandSplitter`, if a file `foo1/bar.o` already exists in the `.a`, then a second `foo2/bar.o` would overwrite the first leading to undefined symbols.

This only happens when *updating* an existing archive (e.g when the `ar` command is long enough to require splitting).

The fix is to pass in `q` instead of `r` to skip indexing and prevent overwriting the first `foo1/bar.o`

In addition, on GNU `ar`, omitting `s` is needed as well.

Test Plan: CI

Reviewed By: bhamiltoncx

fb-gh-sync-id: 24a2ba8