Easy: Fix Python 2.7 check in ng.py
Summary: We needed to check if `memoryview()` was present to support
Python 2.6 in the Nailgun Python client.

Evidently `'foo' in dir(__builtin__)` works in the Python REPL but not
from a module.

This changes the logic to work both from the Python REPL and from
a module.

Test Plan: `buck build buck`, kill buckd PID, confirm stack trace includes
`recv_into_memoryview()` and not `recv_to_buffer_with_copy()`.

Reviewed By: sdwilsh, sid0

fb-gh-sync-id: 8692117