Decrease memory churn of input based rule keys
Summary: Don't build sets of `deps` and `inputs` in
`InputBasedRuleKeyBuilderFactory` until it is necessary. Those sets
can get pretty big and instantiating all of them takes up a lot of
memory. For a big graph with a lot of cxx rules, this change brings
down the total size of allocations related to
`InputBasedRuleKeyBuilderFactory` from 91% to 17%.

Test Plan: `buck test`.

Reviewed By: Coneko

fb-gh-sync-id: f6fd01f