Untracked files now affect get_dirty_buck_version
Before, the presence of untracked files caused Buck to return a dirty
version, but those untracked files were not used in calculating the
dirty version. This meant that creating the first untracked file
caused the buck daemon to restart, but any subsequent modifications to
any untracked files did not. Now, any change to files tracked or
untracked (but not ignored) will affect the dirty version.