thrift: use exported deps for generated C/C++ rules
Summary: The recent changes to C/C++ linking means we're stricter about how
deps are specified.  In the case of thrift, there's a lot of template
code which requires using exported deps.  Since we don't have a way
via the UI for the user to control this (and since we're likely moving
the thrift support into macros which means it's probably not worth
solving this yet) just have the generated C/C++ rule use exported deps.

Test Plan: Updated tests to check for exported deps.

Reviewed By: nayrmi

fb-gh-sync-id: 7b94de0