[Docs] Document the `main_module` property on `python_binary`
Summary: When reading the `python_binary` code I saw mention of a property which was undocumented, and that using the old property would give a deprecation warning. This seemed unintentional, so in this diff I document the new property and update the documentation of the old property to mark it as deprecated.

Test Plan: `./docs/soyweb-local.sh`, view the `python_binary` docs, see the new property listed.

Reviewed By: sdwilsh

fb-gh-sync-id: e6b625e