zip_rule: don't zip entries as symlinks
Summary:
Previously, we'd setup a symlink tree to sandbox the files we'd
zip up, which meant we'd zip them up as symlinks rather than the
actual files.  This diff switches to copying into the sandbox dir
instead.

On another note, our symlink zipping is broken, as we set the symlink
attribute, but then zip up the file contents rather than the link
target.

Closes #542

Test Plan: updated tests to verify perms

Reviewed By: bhamiltoncx

fb-gh-sync-id: 7255487