[Docs] Document `buck run`
Summary:
This command seemed like it was originally supposed to be an implementation detail of `buck install`, but as we've grown to building more and more command line utilities as well its becoming increasingly important for everday use. It should be documented so people know they can use it.

The documentation itself isn't really that useful (since it doesnt really have any CLI arguments of its own) but I think its important for people to know this command exists so that they can use it when using Buck for python/rust/whatever.

Test Plan: `./docs/soyweb-local.sh`, see `run` as an option in the Commands nav. Go to the page for `buck run` and see that it renders correctly.

Reviewed By: sdwilsh

fb-gh-sync-id: d409571