Fix TimeFormatTest failing in non-US English locale
Summary:
This is the last test which fails in non-US English locales.

The only places it's used are from `TestResultSummary` and `TestCaseSummary`.

Refactoring `TestResultSummary` to support multiple locales is pretty painful, so I hard-coded
that class to US English.

Refactoring `TestCaseSummary` isn't as bad, so I'll separately send out a diff to fix that. Temporarily, this diff hard-codes it to US English as well.

Partial fix for https://github.com/facebook/buck/issues/557

Test Plan: `ant java-test -Dtest.class=TimeFormatTest`

Reviewed By: k21

fb-gh-sync-id: 78752f4