Easy: Fix Turkish 'I' bug in AdbHelper
Summary:
`ant forbidden-checks` found this one!

`AdbHelper.deviceUninstallPackage()` is a wrapper for
the Android utility `pm uninstall`.

It tried to do a case-insensitive match on the string `failure`
on the output of the utility by calling `String.toLowerCase()`.

Unfortunately, the output of that method is locale-specific and
depends on the default locale of the system.

In particular, in Turkish, the string "FAILURE" will lower-case
to the string "ı", and the check will not pass.

This fixes the logic by passing in `Locale.US` to ensure we
don't use the default locale.

Test Plan: `ant forbidden-checks` no longer finds this error

Reviewed By: ryu2

fb-gh-sync-id: 7ff3837