New utility MoreFiles.deleteRecursivelyWithOptions()
Summary:
Upon upgrade, Buck currently blocks for a long time at
startup cleaning up the `buck-out` directory synchronously.

This work can be done asynchronously, so I plan to introduce a
`buck-out/.trash` directory to hold the garbage which we clean
up in the background.

I'd like to delete the *contents* of this directory but not
the directory itself, so this diff adds options support to
`MoreFiles.deleteRecursively()` via a new method
`MoreFiles.deleteRecursivelyWithOptions()`.

Test Plan: Unit tests updated. `ant java-test -Dtest.class=MoreFilesTest`

Reviewed By: k21

fb-gh-sync-id: 4659d00