Don't hang for builds with >3000 targets.
Summary:
Futures.transform uses a directExecutor. Doing this enough times
will cause a StackOverflowError due to the nested execute calls.
Because Futures execution works from the bottom, the bottom-most
N (in our case it was 3000) Futures will be able to complete, but
the top-most ones will be left hanging.
The solution is to not wrap as much.

Test Plan: CI

Reviewed By: shs96c

fb-gh-sync-id: 8989e23