Perf: Use WatchmanClient from WatchmanWatcher
Summary:
This ports `WatchmanWatcher` to re-use the `WatchmanClient` created in `Main`.

This way we don't have to spawn off CLI processes every time we want
to talk to Watchman: we'll just re-use the local (Unix domain /
Windows named pipe) socket we previously connected.

In addition, we can use the more efficient BSER protocol instead of
JSON. (This actually does make a difference when there are tens or
hundreds of thousands of changes, which happens in a rebase of a large
repository).

Because I deleted the JSON serialization logic, I removed a test that
only tested JSON escaping.

Test Plan: Updated `WatchmanWatcherTest`. `ant java-test -Dtest.class=WatchmanWatcherTest`

Reviewed By: shs96c

fb-gh-sync-id: 3bc03cc