Set an appropriate concurrency level for the `OptimisticLoadingCache`
Summary:
The backing store of the `OptimisticLoadingCache` is a
`ConcurrentHashMap`.  It has a constructor that takes the expected
concurrency level, which we know to be one when parsing serially,
and the number of parsing threads when parsing in parallel.  This
change sets the concurrency level to the number of threads, and also
sets a bigger default size (256 compared to the default of 16) since
resizing operations are expensive.

Sun's documentation on this says:

> Ideally, you should choose a value to accommodate as many threads
> as will ever concurrently modify the table.

Test Plan: CI

Reviewed By: andrewjcg

fb-gh-sync-id: 0b348b3