Add infrastructure to enable the parallel parsing of build files
Summary:
This adds the infrastrucutre needed to allow parallel parsing of
build files to obtain all `TargetNode`s in them.  This infrastructure
should be the last major bits we need to enable parallel parsing
in the rest of `Parser` (with the exception of `buck project` with no
slice).

Test Plan: CI

Reviewed By: marcinkosiba

fb-gh-sync-id: 125f593