python: fix extension locations in merged link strategy
Summary:
Previously, using omnibus linking grouped the linked roots with all
other native libraries namespaced using the shared library name.
However, in languages such as Python, the roots may be C/C++ extensions,
and therefore need to be loaded via the module search path via their
modules names.

This makes the omnibus linker return the generated root libraries
separately from the others, so that the caller can special case their
layout and handling.  This is used in the Python code, so that C/C++
extensions which are roots in the link get properly placed in the module
tree.

Test Plan: added test

Reviewed By: beefon

fb-gh-sync-id: 82bed8f