Add buildfile operator to buck query
Summary:
Add a 'buildfile' operator to 'buck query'. This operator finds the build files that define the given targets.
The 'owner' operator can be combined with this one in order to know the build file that defines the target that owns a source file.

The motivation behind this commit was a conversation with bolinfest a while ago, that it would be nice to jump to the BUCK file that defines a given source file. Sample usage:
$ buck query "buildfile(owner('src/path/to/file'))"
Closes https://github.com/facebook/buck/pull/590

Test Plan:

- Added tests
- CI
- Published docs locally with $ bash docs/soyweb-local.sh

Reviewed By: sdwilsh

fb-gh-sync-id: 035592e