Fix Turkish I crash in AndroidBinaryDescription
Summary:
If Buck was run on a system in Turkish locale,
creating an `android_binary()` rule with `package_type = "instrumented"`
would crash with `IllegalArgumentException`.

This was because we were using `String.toUpperCase()` to convert the
`String` to an `Enum` without passing in a `Locale`. This would
default to the system locale, which in the case of Turkish has
special rules to map `i` to `\u0310`, causing a crash in
`Enum.valueOf()`.

This fixes the bug and adds a test.

Test Plan:
Test added. Reproduced without fix by running
`ant java-test -Dtest.class=AndroidBinaryDescriptionTest`. Confirmed
crash:

    [junit] Testcase: turkishCaseRulesDoNotCrashConstructor took 0.013 sec
    [junit]     Caused an ERROR
    [junit] No enum constant com.facebook.buck.android.AndroidBinary.PackageType.İNSTRUMENTED
    [junit] java.lang.IllegalArgumentException: No enum constant com.facebook.buck.android.AndroidBinary.PackageType.İNSTRUMENTED
    [junit]     at java.lang.Enum.valueOf(Enum.java:238)
    [junit]     at com.facebook.buck.android.AndroidBinary$PackageType.valueOf(AndroidBinary.java:1)
    [junit]     at com.facebook.buck.android.AndroidBinaryDescription.getPackageType(AndroidBinaryDescription.java:293)
    [junit]     at com.facebook.buck.android.AndroidBinaryDescription.createBuildRule(AndroidBinaryDescription.java:170)
    [junit]     at com.facebook.buck.android.AndroidBinaryDescription.createBuildRule(AndroidBinaryDescription.java:1)
    [junit]     at com.facebook.buck.rules.AbstractNodeBuilder.build(AbstractNodeBuilder.java:103)
    [junit]     at com.facebook.buck.android.AndroidBinaryDescriptionTest.turkishCaseRulesDoNotCrashConstructor(AndroidBinaryDescriptionTest.java:97)
    [junit]

Applied fix, re-ran test, no more crash.

Reviewed By: ryu2

fb-gh-sync-id: 240b059