Fix `ProvisioningProfileCopyStepTest` to copy in files
Summary:
This test uses a `FakeProjectFilesystem`, but then depends on files being accessed
from the disk.  In a future change, I implement the `copy` method on
`FakeProjectFilesystem`, which breaks this test.

This change copies the needed files into the filesystem so that the test continues
to pass.

Test Plan: CI

Reviewed By: ryu2

fb-gh-sync-id: 50b1f70