Added directory zipping to File Bundler
Summary:
This diff makes it so that BUCK can also handle zipping files in a directory.
If it is a directory from buck-out/gen, and it has the same base path, then it's going to keep the hierarchy, otherwise, it's just going to add the file to the zip/jar.

I need this in order to make the buck plugin build with Buck and zip the generated files.

Test Plan: I tested it with buck and the buck plugin for generated files.

Reviewed By: marcinkosiba

fb-gh-sync-id: 1a073a0