Do not implicitly add -Xlinker flag for apple platforms when generating Xcode projects
Summary:
Right now, when we specify custom `linker_flags` we add `-Xlinker` in-between the flags in Xcode project generation, but not in native builds.  This results in needing two different ways to specify the flags for project and build.

Remove the implicit `-Xlinker` so that the flags can be unified.

Test Plan: CI

Reviewed By: andrewjcg

fb-gh-sync-id: fad97e1