parser: add mechanism to read `.buckconfig` values from build files
Summary:
Adds a new `read_config()` function for use inside BUCK files, providing
a way to read `.buckconfig` entries at parse time.  The idea is this provides
a managed way thread in information without resorting to things like
environment variables.  When a config setting is read, we record it and
remember it along with the cached build file, and make sure we invalidate it
when it's changed, added, or removed.

Test Plan: added unittests

Reviewed By: Coneko

fb-gh-sync-id: e3f2ce2