Merge different code paths for preprocessor delegates.
Summary:
Preprocessor delegates have two different code paths because the
code was originally for combined and separate preprocessing.
However when piped mode was added it used the path for combined.
This means the preprocessing for piped is missing some flags.
Since duplicating the flags is not a big issue, and we plan on
reducing the preprocessing modes to only one anyway, we can
merge the code paths.

Test Plan: CI

Reviewed By: k21

fb-gh-sync-id: 5e514df