Make logger to not block threads when its publishing logs to disk
Summary:
Java Logger at some point need to publish the logs to disk. Since we have an instance of a logger per class,
it may block the thread which logs into it if the logger's java.Logger inctance will decide to dump it's logs
to the disk.
By wrapping the log call into executor we make sure only designated thread will be blocked.

Test Plan:
CI
Profile and make sure build threads are not blocked from time to time by the Logger

Reviewed By: bhamiltoncx

fb-gh-sync-id: 8ebdd44