[buck] Add simple MacroMatchResult class
Summary: This diff replaces the ##Pair<String, String>## returned by ##MacroFinder.findAll(...)## with a simple Immutable called ##MacroMatchResult##. In addition to the ##macroType## and ##macroInput## that was previously contained in the ##Pair##, this class also contains the start and end indices of where this macro was found in the overall string.

Test Plan: Updated MacroFinderTest and ran it.

Reviewed By: Coneko

fb-gh-sync-id: add88f1